refactor: dedup value marshalling into backend-common
The CEF and webview backends each hand-duplicated the same ~35 value_*
marshalling functions (cef stored CefValue, webview stored wef::Value).
Extract one canonical implementation:

- backend-common/wef_value.{h,cc}: the wef::Value type + the value_*
  functions, installed via wef_register_value_api(). Both backends link it.
- webview: drop its copy, register the shared impl (already used wef::Value).
- cef: store wef::Value too; convert CefValue<->wef::Value only at the
  renderer<->browser IPC boundary (CefValueToWef / WefToCefValue). Decodes
  {"__callback__":id} dicts to wef::Value::Callback, matching webview.

Net ~600 lines of duplication removed.
