+++ .gitignore
@@ -7,3 +7,4 @@ build/
 */build*/
 webview/packages/
 /book
+result-*
