+++ cef/src/runtime_loader.cc
@@ -415,282 +415,133 @@ static void Backend_PostUiTask(void* data, void (*task)(void*),
   }
 }
 
-// --- Value accessors ---
+// --- CefValue <-> wef::Value conversion (IPC boundary only) ---
+//
+// Values cross the renderer<->browser process boundary as CefValue trees, but
+// the shared marshalling layer operates on wef::Value. Convert once on the way
+// in (incoming JS args / eval results) and once on the way out (responses,
+// callback args). A JS function is encoded by the renderer as a dictionary
+// {"__callback__": "<id>"}; decode it to wef::Value::Callback so that
+// value_is_callback works, matching the webview backend.
 
-static bool Backend_ValueIsNull(wef_value_t* val) {
-  if (!val || !val->value)
-    return true;
-  return val->value->GetType() == VTYPE_NULL;
-}
-
-static bool Backend_ValueIsBool(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_BOOL;
-}
-
-static bool Backend_ValueIsInt(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_INT;
-}
-
-static bool Backend_ValueIsDouble(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_DOUBLE;
-}
-
-static bool Backend_ValueIsString(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_STRING;
-}
-
-static bool Backend_ValueIsList(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_LIST;
-}
-
-static bool Backend_ValueIsDict(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_DICTIONARY;
-}
-
-static bool Backend_ValueIsBinary(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetType() == VTYPE_BINARY;
-}
-
-static bool Backend_ValueIsCallback(wef_value_t* val) {
-  if (!val)
-    return false;
-  return val->is_callback;
-}
-
-static bool Backend_ValueGetBool(wef_value_t* val) {
-  if (!val || !val->value)
-    return false;
-  return val->value->GetBool();
-}
-
-static int Backend_ValueGetInt(wef_value_t* val) {
-  if (!val || !val->value)
-    return 0;
-  return val->value->GetInt();
-}
-
-static double Backend_ValueGetDouble(wef_value_t* val) {
-  if (!val || !val->value)
-    return 0.0;
-  return val->value->GetDouble();
-}
-
-static char* Backend_ValueGetString(wef_value_t* val, size_t* len_out) {
-  if (!val || !val->value) {
-    if (len_out)
-      *len_out = 0;
-    return nullptr;
-  }
-  std::string str = val->value->GetString().ToString();
-  if (len_out)
-    *len_out = str.size();
-  char* result = static_cast<char*>(malloc(str.size() + 1));
-  if (result) {
-    memcpy(result, str.c_str(), str.size() + 1);
-  }
-  return result;
-}
-
-static void Backend_ValueFreeString(char* str) {
-  free(str);
-}
-
-static size_t Backend_ValueListSize(wef_value_t* val) {
-  if (!val || !val->value || val->value->GetType() != VTYPE_LIST)
-    return 0;
-  return val->value->GetList()->GetSize();
-}
-
-static wef_value_t* Backend_ValueListGet(wef_value_t* val, size_t index) {
-  if (!val || !val->value || val->value->GetType() != VTYPE_LIST)
-    return nullptr;
-  CefRefPtr<CefListValue> list = val->value->GetList();
-  if (index >= list->GetSize())
-    return nullptr;
-  return new wef_value(list->GetValue(index));
-}
-
-static wef_value_t* Backend_ValueDictGet(wef_value_t* dict, const char* key) {
-  if (!dict || !dict->value || dict->value->GetType() != VTYPE_DICTIONARY ||
-      !key)
-    return nullptr;
-  CefRefPtr<CefDictionaryValue> d = dict->value->GetDictionary();
-  if (!d->HasKey(key))
-    return nullptr;
-  return new wef_value(d->GetValue(key));
-}
-
-static bool Backend_ValueDictHas(wef_value_t* dict, const char* key) {
-  if (!dict || !dict->value || dict->value->GetType() != VTYPE_DICTIONARY ||
-      !key)
-    return false;
-  return dict->value->GetDictionary()->HasKey(key);
-}
-
-static size_t Backend_ValueDictSize(wef_value_t* dict) {
-  if (!dict || !dict->value || dict->value->GetType() != VTYPE_DICTIONARY)
-    return 0;
-  return dict->value->GetDictionary()->GetSize();
-}
-
-static char** Backend_ValueDictKeys(wef_value_t* dict, size_t* count_out) {
-  if (!dict || !dict->value || dict->value->GetType() != VTYPE_DICTIONARY) {
-    if (count_out)
-      *count_out = 0;
-    return nullptr;
-  }
-  CefRefPtr<CefDictionaryValue> d = dict->value->GetDictionary();
-  CefDictionaryValue::KeyList keys;
-  d->GetKeys(keys);
-
-  if (count_out)
-    *count_out = keys.size();
-  if (keys.empty())
-    return nullptr;
-
-  char** result = static_cast<char**>(malloc(sizeof(char*) * keys.size()));
-  if (!result) {
-    if (count_out)
-      *count_out = 0;
-    return nullptr;
-  }
-  for (size_t i = 0; i < keys.size(); ++i) {
-    std::string key = keys[i].ToString();
-    result[i] = static_cast<char*>(malloc(key.size() + 1));
-    if (!result[i]) {
-      // Roll back rather than hand back a half-built array.
-      for (size_t j = 0; j < i; ++j)
-        free(result[j]);
-      free(result);
-      if (count_out)
-        *count_out = 0;
-      return nullptr;
+static wef::ValuePtr CefValueToWef(CefRefPtr<CefValue> v) {
+  if (!v)
+    return wef::Value::Null();
+  switch (v->GetType()) {
+    case VTYPE_BOOL:
+      return wef::Value::Bool(v->GetBool());
+    case VTYPE_INT:
+      return wef::Value::Int(v->GetInt());
+    case VTYPE_DOUBLE:
+      return wef::Value::Double(v->GetDouble());
+    case VTYPE_STRING:
+      return wef::Value::String(v->GetString().ToString());
+    case VTYPE_BINARY: {
+      CefRefPtr<CefBinaryValue> bin = v->GetBinary();
+      std::vector<uint8_t> buf(bin->GetSize());
+      if (!buf.empty())
+        bin->GetData(buf.data(), buf.size(), 0);
+      return wef::Value::Binary(buf.data(), buf.size());
     }
-    memcpy(result[i], key.c_str(), key.size() + 1);
+    case VTYPE_LIST: {
+      CefRefPtr<CefListValue> list = v->GetList();
+      auto out = wef::Value::List();
+      for (size_t i = 0; i < list->GetSize(); ++i) {
+        out->GetList().push_back(CefValueToWef(list->GetValue(i)));
+      }
+      return out;
+    }
+    case VTYPE_DICTIONARY: {
+      CefRefPtr<CefDictionaryValue> dict = v->GetDictionary();
+      if (dict->HasKey("__callback__")) {
+        // Renderer-supplied; don't let a malformed id abort the process.
+        std::string id = dict->GetString("__callback__").ToString();
+        try {
+          return wef::Value::Callback(static_cast<uint64_t>(std::stoull(id)));
+        } catch (const std::exception&) {
+          return wef::Value::Null();
+        }
+      }
+      auto out = wef::Value::Dict();
+      CefDictionaryValue::KeyList keys;
+      dict->GetKeys(keys);
+      for (const auto& key : keys) {
+        out->GetDict()[key.ToString()] = CefValueToWef(dict->GetValue(key));
+      }
+      return out;
+    }
+    case VTYPE_NULL:
+    case VTYPE_INVALID:
+    default:
+      return wef::Value::Null();
   }
-  return result;
 }
 
-static void Backend_ValueFreeKeys(char** keys, size_t count) {
-  if (!keys)
-    return;
-  for (size_t i = 0; i < count; ++i) {
-    free(keys[i]);
+static CefRefPtr<CefValue> WefToCefValue(const wef::ValuePtr& v) {
+  CefRefPtr<CefValue> out = CefValue::Create();
+  if (!v) {
+    out->SetNull();
+    return out;
   }
-  free(keys);
-}
-
-static const void* Backend_ValueGetBinary(wef_value_t* val, size_t* len_out) {
-  if (!val || !val->value || val->value->GetType() != VTYPE_BINARY) {
-    if (len_out)
-      *len_out = 0;
-    return nullptr;
+  switch (v->type) {
+    case wef::ValueType::Bool:
+      out->SetBool(v->GetBool());
+      break;
+    case wef::ValueType::Int:
+      out->SetInt(v->GetInt());
+      break;
+    case wef::ValueType::Double:
+      out->SetDouble(v->GetDouble());
+      break;
+    case wef::ValueType::String:
+      out->SetString(v->GetString());
+      break;
+    case wef::ValueType::Binary: {
+      const auto& bin = v->GetBinary();
+      out->SetBinary(CefBinaryValue::Create(bin.data.data(), bin.data.size()));
+      break;
+    }
+    case wef::ValueType::List: {
+      CefRefPtr<CefListValue> list = CefListValue::Create();
+      const auto& items = v->GetList();
+      for (size_t i = 0; i < items.size(); ++i) {
+        list->SetValue(i, WefToCefValue(items[i]));
+      }
+      out->SetList(list);
+      break;
+    }
+    case wef::ValueType::Dict: {
+      CefRefPtr<CefDictionaryValue> dict = CefDictionaryValue::Create();
+      for (const auto& pair : v->GetDict()) {
+        dict->SetValue(pair.first, WefToCefValue(pair.second));
+      }
+      out->SetDictionary(dict);
+      break;
+    }
+    case wef::ValueType::Callback: {
+      CefRefPtr<CefDictionaryValue> dict = CefDictionaryValue::Create();
+      dict->SetString("__callback__", std::to_string(v->GetCallbackId()));
+      out->SetDictionary(dict);
+      break;
+    }
+    case wef::ValueType::Null:
+    default:
+      out->SetNull();
+      break;
   }
-  CefRefPtr<CefBinaryValue> binary = val->value->GetBinary();
-  if (len_out)
-    *len_out = binary->GetSize();
-  return binary->GetRawData();
+  return out;
 }
 
-static uint64_t Backend_ValueGetCallbackId(wef_value_t* val) {
-  if (!val || !val->is_callback)
-    return 0;
-  return val->callback_id;
-}
-
-static wef_value_t* Backend_ValueNull(void*) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetNull();
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueBool(void*, bool v) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetBool(v);
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueInt(void*, int v) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetInt(v);
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueDouble(void*, double v) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetDouble(v);
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueString(void*, const char* v) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetString(v ? v : "");
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueList(void*) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetList(CefListValue::Create());
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueDict(void*) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  val->SetDictionary(CefDictionaryValue::Create());
-  return new wef_value(val);
-}
-
-static wef_value_t* Backend_ValueBinary(void*, const void* data, size_t len) {
-  CefRefPtr<CefValue> val = CefValue::Create();
-  CefRefPtr<CefBinaryValue> binary = CefBinaryValue::Create(data, len);
-  val->SetBinary(binary);
-  return new wef_value(val);
-}
-
-static bool Backend_ValueListAppend(wef_value_t* list, wef_value_t* val) {
-  if (!list || !list->value || list->value->GetType() != VTYPE_LIST)
-    return false;
-  if (!val || !val->value)
-    return false;
-  CefRefPtr<CefListValue> l = list->value->GetList();
-  size_t index = l->GetSize();
-  return l->SetValue(index, val->value);
-}
-
-static bool Backend_ValueListSet(wef_value_t* list, size_t index,
-                                 wef_value_t* val) {
-  if (!list || !list->value || list->value->GetType() != VTYPE_LIST)
-    return false;
-  if (!val || !val->value)
-    return false;
-  return list->value->GetList()->SetValue(index, val->value);
-}
-
-static bool Backend_ValueDictSet(wef_value_t* dict, const char* key,
-                                 wef_value_t* val) {
-  if (!dict || !dict->value || dict->value->GetType() != VTYPE_DICTIONARY)
-    return false;
-  if (!key || !val || !val->value)
-    return false;
-  return dict->value->GetDictionary()->SetValue(key, val->value);
-}
-
-static void Backend_ValueFree(wef_value_t* val) {
-  delete val;
+// Build a CefListValue from a wef list value (empty list otherwise).
+static CefRefPtr<CefListValue> WefListToCef(const wef::ValuePtr& v) {
+  CefRefPtr<CefListValue> list = CefListValue::Create();
+  if (v && v->IsList()) {
+    const auto& items = v->GetList();
+    for (size_t i = 0; i < items.size(); ++i) {
+      list->SetValue(i, WefToCefValue(items[i]));
+    }
+  }
+  return list;
 }
 
 // --- JS call/callback handling ---
@@ -716,7 +567,7 @@ static void Backend_JsCallRespond(void* data, uint64_t call_id,
   args->SetDouble(0, static_cast<double>(call_id));
 
   if (result && result->value) {
-    args->SetValue(1, result->value);
+    args->SetValue(1, WefToCefValue(result->value));
   } else {
     CefRefPtr<CefValue> null_val = CefValue::Create();
     null_val->SetNull();
@@ -724,7 +575,7 @@ static void Backend_JsCallRespond(void* data, uint64_t call_id,
   }
 
   if (error && error->value) {
-    args->SetValue(2, error->value);
+    args->SetValue(2, WefToCefValue(error->value));
   } else {
     CefRefPtr<CefValue> null_val = CefValue::Create();
     null_val->SetNull();
@@ -747,11 +598,7 @@ static void Backend_InvokeJsCallback(void* data, uint64_t callback_id,
   CefRefPtr<CefListValue> msgArgs = msg->GetArgumentList();
   msgArgs->SetDouble(0, static_cast<double>(callback_id));
 
-  if (args && args->value && args->value->GetType() == VTYPE_LIST) {
-    msgArgs->SetList(1, args->value->GetList());
-  } else {
-    msgArgs->SetList(1, CefListValue::Create());
-  }
+  msgArgs->SetList(1, WefListToCef(args ? args->value : nullptr));
 
   loader->ForEachBrowser([&msg](CefRefPtr<CefBrowser> browser) {
     CefPostTask(
@@ -1394,44 +1241,7 @@ void RuntimeLoader::InitializeBackendApi() {
   backend_api_.focus = Backend_Focus;
   backend_api_.post_ui_task = Backend_PostUiTask;
 
-  backend_api_.value_is_null = Backend_ValueIsNull;
-  backend_api_.value_is_bool = Backend_ValueIsBool;
-  backend_api_.value_is_int = Backend_ValueIsInt;
-  backend_api_.value_is_double = Backend_ValueIsDouble;
-  backend_api_.value_is_string = Backend_ValueIsString;
-  backend_api_.value_is_list = Backend_ValueIsList;
-  backend_api_.value_is_dict = Backend_ValueIsDict;
-  backend_api_.value_is_binary = Backend_ValueIsBinary;
-  backend_api_.value_is_callback = Backend_ValueIsCallback;
-
-  backend_api_.value_get_bool = Backend_ValueGetBool;
-  backend_api_.value_get_int = Backend_ValueGetInt;
-  backend_api_.value_get_double = Backend_ValueGetDouble;
-  backend_api_.value_get_string = Backend_ValueGetString;
-  backend_api_.value_free_string = Backend_ValueFreeString;
-  backend_api_.value_list_size = Backend_ValueListSize;
-  backend_api_.value_list_get = Backend_ValueListGet;
-  backend_api_.value_dict_get = Backend_ValueDictGet;
-  backend_api_.value_dict_has = Backend_ValueDictHas;
-  backend_api_.value_dict_size = Backend_ValueDictSize;
-  backend_api_.value_dict_keys = Backend_ValueDictKeys;
-  backend_api_.value_free_keys = Backend_ValueFreeKeys;
-  backend_api_.value_get_binary = Backend_ValueGetBinary;
-  backend_api_.value_get_callback_id = Backend_ValueGetCallbackId;
-
-  backend_api_.value_null = Backend_ValueNull;
-  backend_api_.value_bool = Backend_ValueBool;
-  backend_api_.value_int = Backend_ValueInt;
-  backend_api_.value_double = Backend_ValueDouble;
-  backend_api_.value_string = Backend_ValueString;
-  backend_api_.value_list = Backend_ValueList;
-  backend_api_.value_dict = Backend_ValueDict;
-  backend_api_.value_binary = Backend_ValueBinary;
-
-  backend_api_.value_list_append = Backend_ValueListAppend;
-  backend_api_.value_list_set = Backend_ValueListSet;
-  backend_api_.value_dict_set = Backend_ValueDictSet;
-  backend_api_.value_free = Backend_ValueFree;
+  wef_register_value_api(&backend_api_);
 
   backend_api_.set_js_call_handler = Backend_SetJsCallHandler;
   backend_api_.js_call_respond = Backend_JsCallRespond;
@@ -1729,12 +1539,10 @@ void RuntimeLoader::HandleEvalResult(uint64_t eval_id,
   }
 
   if (!error.empty()) {
-    CefRefPtr<CefValue> errValue = CefValue::Create();
-    errValue->SetString(error);
-    wef_value errWef(errValue);
+    wef_value errWef(wef::Value::String(error));
     eval.callback(nullptr, &errWef, eval.callback_data);
   } else if (result && result->GetType() != VTYPE_NULL) {
-    wef_value resultWef(result);
+    wef_value resultWef(CefValueToWef(result));
     eval.callback(&resultWef, nullptr, eval.callback_data);
   } else {
     eval.callback(nullptr, nullptr, eval.callback_data);
@@ -1786,13 +1594,12 @@ void RuntimeLoader::PollPendingJsCalls() {
     if (handler) {
       CefRefPtr<CefValue> argsValue = CefValue::Create();
       argsValue->SetList(call.args);
-      wef_value_t* argsWrapper = new wef_value(argsValue);
+      wef_value_t* argsWrapper = new wef_value(CefValueToWef(argsValue));
       handler(user_data, call.window_id, call.call_id, call.method_path.c_str(),
               argsWrapper);
     } else {
-      CefRefPtr<CefValue> errVal = CefValue::Create();
-      errVal->SetString("No JS call handler registered");
-      wef_value_t errWrapper(errVal);
+      wef_value_t errWrapper(
+          wef::Value::String("No JS call handler registered"));
       Backend_JsCallRespond(this, call.call_id, nullptr, &errWrapper);
     }
   }
